requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
accelerate=0.32.1=pyhd8ed1ab_0
aiohttp=3.9.5=py39h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=16.1.0=h7cd61ee_0
async-timeout=4.0.3=py39haa95532_0
attrs=23.1.0=py39haa95532_0
aws-c-auth=0.6.19=h2bbff1b_0
aws-c-cal=0.5.20=h2bbff1b_0
aws-c-common=0.8.5=h2bbff1b_0
aws-c-compression=0.2.16=h2bbff1b_0
aws-c-event-stream=0.2.15=hd77b12b_0
aws-c-http=0.6.25=h2bbff1b_0
aws-c-io=0.13.10=h2bbff1b_0
aws-c-mqtt=0.7.13=h2bbff1b_0
aws-c-s3=0.1.51=h2bbff1b_0
aws-c-sdkutils=0.1.6=h2bbff1b_0
aws-checksums=0.1.13=h2bbff1b_0
aws-crt-cpp=0.18.16=hd77b12b_0
aws-sdk-cpp=1.10.55=hd77b12b_0
blas=1.0=mkl
boost-cpp=1.82.0=h59b6b97_2
bottleneck=1.3.7=py39h9128911_0
brotli-python=1.0.9=py39hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2024.7.2=haa95532_0
catalogue=2.0.10=py39haa95532_0
certifi=2024.7.4=py39haa95532_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py39haa95532_0
colorama=0.4.6=py39haa95532_0
confection=0.1.4=py39h9909e9c_0
cuda-cccl=12.5.39=0
cuda-cccl_win-64=12.5.39=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-opencl-dev=12.5.39=0
cuda-profiler-api=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cymem=2.0.6=py39hd77b12b_0
cython-blis=0.7.9=py39h080aedc_0
datasets=2.19.1=py39haa95532_0
dill=0.3.8=py39haa95532_0
filelock=3.13.1=py39haa95532_0
frozenlist=1.4.0=py39h2bbff1b_0
fsspec=2024.3.1=py39haa95532_0
gflags=2.2.2=hd77b12b_1
glog=0.5.0=hd77b12b_1
gmpy2=2.1.2=py39h7f96b67_0
huggingface_hub=0.23.5=pyhd8ed1ab_0
idna=3.7=py39haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.4=py39haa95532_0
langcodes=3.3.0=pyhd3eb1b0_0
libabseil=20240116.2=cxx17_h5da7b33_0
libboost=1.82.0=h3399ecb_2
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.6.82=0
libcurand-dev=10.3.6.82=0
libcurl=8.7.1=h86230a5_0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libevent=2.1.12=h56d1f94_1
libgrpc=1.62.2=hf25190f_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libprotobuf=4.25.3=hf2fb9eb_0
libssh2=1.11.0=h291bd65_0
libthrift=0.15.0=h4364b78_2
libuv=1.48.0=h827c3e9_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py39h2bbff1b_0
mkl=2021.4.0=h0e2418a_729
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py39haa95532_0
multidict=6.0.4=py39h2bbff1b_0
multiprocess=0.70.15=py39haa95532_0
murmurhash=1.0.7=py39hd77b12b_0
networkx=3.2.1=py39haa95532_0
numexpr=2.8.4=py39h5b0cc5e_0
numpy=1.24.3=py39hf95b240_0
numpy-base=1.24.3=py39h005ec55_0
openssl=3.0.14=h827c3e9_0
orc=2.0.1=hd8d391b_0
packaging=24.1=py39haa95532_0
pandas=2.2.2=py39h5da7b33_0
pathy=0.10.3=py39haa95532_0
pip=24.0=py39haa95532_0
preshed=3.0.6=py39h6c2663c_0
psutil=5.9.0=py39h2bbff1b_0
pyarrow=16.1.0=py39hc64d2fc_0
pydantic=1.10.12=py39h2bbff1b_1
pysocks=1.7.1=py39haa95532_0
python=3.9.19=h1aa4202_1
python-dateutil=2.9.0post0=py39haa95532_2
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py39h2bbff1b_1
pytorch=2.3.1=py3.9_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytz=2024.1=py39haa95532_0
pyyaml=6.0.1=py39h2bbff1b_0
re2=2022.04.01=hd77b12b_0
regex=2023.10.3=py39h2bbff1b_0
requests=2.32.3=py39haa95532_0
safetensors=0.4.2=py39h9e6c545_1
setuptools=69.5.1=py39haa95532_0
shellingham=1.5.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=py39haa95532_0
snappy=1.1.10=h6c2663c_1
spacy=3.5.3=py39hef0f399_0
spacy-legacy=3.0.12=py39haa95532_0
spacy-loggers=1.0.4=py39haa95532_0
spacy-model-en_core_web_lg=3.5.0=pyhd3eb1b0_1
sqlite=3.45.3=h2bbff1b_0
srsly=2.4.8=py39hd77b12b_1
sympy=1.12=py39haa95532_0
tbb=2021.8.0=h59b6b97_0
thinc=8.1.10=py39hf497b98_0
tokenizers=0.19.1=py39h7ac788d_0
tqdm=4.66.4=py39h9909e9c_0
transformers=4.42.4=pyhd8ed1ab_0
typer=0.4.1=py39haa95532_0
typing-extensions=4.11.0=py39haa95532_0
typing_extensions=4.11.0=py39haa95532_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=py39haa95532_0
utf8proc=2.6.1=h2bbff1b_1
vc=14.2=h2eaa2aa_4
vs2015_runtime=14.29.30133=h43f2093_4
wasabi=0.9.1=py39haa95532_0
wheel=0.43.0=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
xxhash=0.8.0=h2bbff1b_3
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.9.3=py39h2bbff1b_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
